Frontend: Split view module
This is desired due to the length of the module and the need to extend it in the near future due to the regenerate and NetBox features.
