+++ orthos2/frontend/views/__init__.py
@@ -1,0 +1,10 @@
+"""
+This module contains all Django frontend views. For convenience, they are all reexported.
+"""
+
+from .ajax import *
+from .auth import *
+from .machine import *
+from .machines import *
+from .statistics import *
+from .user import *
