+++ orthos2/frontend/views/ajax.py
